requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
tqdm==4.66.4
pdfminer.six==20231228
google-generativeai==0.3.2
pandas==2.2.2
matplotlib==3.8.4
seaborn==0.13.2
numpy==1.26.4
scipy==1.13.0
adjustText==0.8
pylr2==0.1.0
nltk==3.8.1
networkx==3.5
pydot==4.0.1
scikit-learn==1.4.2